Update and rename 2467.cpp to 2467_1.cpp
